Changed the general aliases to use long options when possible.
The general aliases have been updated to use the long options for
commands now (e.g. `-a` vs `--all`). This is done to make it more clear
what the options do. Also, some minor fixups were added as well. Last,
the semantics of the `ll` and `la` aliases were switched.