Updated the aliases for command wrapped by the readline library.
The aliases for wrapping commands with the readline library has been
generalized. These commands are now specified as a list, and the aliases
are generated on the fly. Additionally, the readline settings were
tweaked to allow for (near) infinite command history, aggressive
elimination of duplicates, and a unique history file for each command
that is wrapped by the library.