Added in setting the 'pipefail' option.
This changes Bash's default behavior so that when executing a pipeline
of commands, the exit status of the pipeline is 0 if all commands
succeed, or the exit status of the rightmost command with a non-zero
exit status. Without this option, the exit status only comes from the
final command in the pipeline, irrespective of the other commands in the
pipeline.