Fixed an issue with how spaces are treated for Windows commands.
Because of the way the run Windows function was setup before, it would
cause any arguments that are quoted (because they contain spaces) to be
improperly handled. This is because the command was fed through `sed`
en-masse. Now, each individual entry (command and each argument) of the
command is fed through `sed` to replace the drive letter. This allows
the spaces in arguments to be preserved, and prevent it from being
parsed as more than one argument.