Added in handling of Unix-style paths for the program path for the WSL.
The `run-windows` function now properly handles Unix-style paths for the
program path (argument 0). If applicable, forward slashes are converted
to back slashes. This is only done for the program path, because windows
programs can typically understand Unix-style paths, and there isn't a
good way to do it generally for arguments. If the program is a command,
or in the PATH, then no conversion is necessary.

Comments in the function were also updated and made more descriptive.