Updated the color settings for Git.
The default Git color settings has now been set to "always", which means
that colors are generated by commands even if they are being piped to
another process. This is useful because these commands are generally not
written to files, so preserving color codes is desired. The exception to
this is Diff, which can be used to create patch files. The settings for
this command are so colors are not generated when being written to file.