Added proper handling of symlinks to the WSL's `run-windows` function.
The `run-windows` function for the WSL now properly handles symbolic
links in paths. For the program (first argument), symbolic links are
resolved (and the path is canonicalized to an absolute path) if the path
has more than one component (e.g. msbuild vs bin/msbuild). For the
program arguments, their paths have symbolic links resolved
unconditionally.

The reason this is necessary is because a common pattern in a WSL setup
is to have symbolic links in the Linux/WSL home directory that link to
paths on the Windows filesystem (typically the Windows user's home
directory). When running Windows programs, these paths must have
symbolic links resolved to their true Windows path in order for programs
to understand them.

This pattern is not common for Git Bash, so updating its `run-windows`
function is not needed.