Fixed an issue with line wrapping and `PS1`.
There was an issue with the Bash prompt variable (`PS1`), that was
causing Bash not to properly wrap long lines in the terminal. The issue
was that '\[' '\]' delimiters were missing from the color codes. This
tells Bash not to count the characters inside towards the line length.
Without this option, Bash was wrapping long lines on the same terminal
line, instead of going to the next line, causing the input to become
mangled and unreadable.