Updated the default verbosity for Git commit.
The default verbosity for the commit command has been updated to 2 in
the Git config. This means that, along with the normal commit template,
a diff of the changes being committed and a diff of the changes not
committed are also placed in the file. This is useful for reviewing what
is being committed.