Added default options for submodules in the Git config.
Two options were added to the Git config for submodules. All commands
that deal with submodules (e.g. pull, fetch, checkout, etc.) will now
recurse into submodules. This deals with the annoying issue that
submodules are not updated when pulling changes or checking out a
previous commit. The other option was to set the number of threads used
to fetch submodules to a reasonable default value. Also, some comments
were updated.