Updated the algorithm used for Git Diff.
The Git config was updated with some additional settings for the diff
command. Diffs now have any whitespace changes highlighted as errors
both in the old and new versions of the file (extraneous whitespace, tab
to space changes, etc.). Additionally, the diffs are now generated with
experimental indent heuristics and the histogram algorithm, which is an
improved version of the patience algorithm. Both of these options serve
to generate better and more intuitive diffs, namely by generating diff
hunks that are code blocks.