Updated the `recent-branches` alias to always use color.
In the most recent version of Git, the For-Each-Ref command has been
updated so that when to use color can be specified. Thus, the
`recent-branches` alias had to be updated to always use color (this was
the default before). This is desirable so that its output can be piped
to the Head and Tail commands.