Added additional settings to the Git config.
The settings for the Merge, Pack, Push, and Rebase commands were updated
in the Git config. The tool for merge was set and the number of threads
to use when packing objects was set to be the maximum useful number of
threads. When pushing to the remote, tags are now also pushed by
default. Last, a warning is printed if any commits are removed during an
interactive rebase (a line is deleted from the to do list).