Added some additional settings to the Git Config.
The settings for the Status, Stash, and Tag commands were updated. When
displaying stashes, they are displayed in patch format by default,
instead of the typical diffstat. The status displayed now also contains
the number of entries currently in the stash list. When displaying tags,
the tags are sorted by committer date (or commit creation date) in
descending order by default.