Added importing of the See module to the Python startup.
The See module is now imported in the Python startup file, so it is
available by default to the user in the interactive session. The See
module provides an enhanced version of the built-in Dir function, which
lists the members of an object. It has some additional features like
the ability to filter the results with wildcard patterns.