Updated the PS2 prompt used by the Python interpreter.
The Python startup file was updated to change the PS2 prompt used by the
interactive session. The PS2 prompt is printed whenever line
continuations occur (such as defining functions). The prompt was updated
to simply be spaces. This was done so that multiline commands could be
properly copied, as they are not stored as a single entry in the
command history.