Changed the behavior of the up and down arrow keys for history.
The Readline initialization file was updated to change the behavior of
the up and down arrow keys. Rather than just linearily going through the
history, the up and down arrow keys now perform a search through history
using the string on the current line. Thus, typing something like `ssh
<up arrow>` will search through all SSH commands in the command history.
When the line is blank, this behaves the same as it originally did.