Added some shell options for changing directories.
Two new shell options for changing directories were added to the Bashrc.
The first option allows users to type a directory as a command, and Bash
will automatically change the directory to that path. The second option
allows users to type a variable name without the dollar sign to the cd
built-in, and the command will interpret the string as a variable name
and expand it to get the directory path to change to.