Changed the Readline shortcuts for killing words.
The Readline initialization file was updated to change the shortcuts for
killing words. The Ctrl-W shortcut was changed to kill a word, instead
of the default Unix rubout function, we treats words as anything
delimited with space. A new Alt-W shortcut was added for killing shell
words. This was done to mirror the different movement shortcuts for
words vs shell words.