Generalized the run Windows functions for WSL and Git Bash.
The run Windows command functions have be split into two functions and
generalized for the WSL and Git Bash variants. The first function runs a
native Windows command in the same command prompt window. The second
function runs a native Windows command in a separate command prompt
window.

The common part of translating the command to be compatible with
invoking in a command prompt shell was moved into a separate helper
function.