Fixed some issues with WSL's run Windows functions.
There were two issues. The first was that the program path's drive
backslash was not being properly escaped (similar to before, since it is
evaluated, double backspaces are required). Second, it is possible to
specify empty strings for the program arguments, so the `--quiet` flag
is needed for the real path command.