Added in local, machine-specific settings to the Python startup file.
The Python startup file now supports including local settings. These are
handled by performing a dynamic wildcard import of the local settings
file. If this file exists, then nothing happens. The main purpose of
local settings is to act as a scratch pad for function aliases used for
development purposes.

To support this new functionality, a new helper function was added that
performs a wildcard import on a Python file given by a path. This was
left as one of the defined functions in the Python startup, as it could
be used in other contexts from the interpreter potentially.