Set the bracketed paste option to only be enabled for Bash.
The Readline initialization file has been updated to disable the
bracketed paste option for all programs except Bash. Bracketed paste
treats the string as-is, and doesn't allow it to perform any editing
commands. For example, newlines are ignored in the pasted content. This
is useful for Bash, where multiline commands that aren't properly
escaped are pasted. However, it is less useful for other programs, such
as Python where you may want to paste multiple separate lines of code.