Fixed an issue with Windows option and the WSL path command.
The Bashrc has been updated to resolve this issue. The WSL path tries to
detect if a specified path is not representable as a Windows path (e.g.
/etc). If it isn't, then it will return an error code. This causes
issues for Windows program options, which typically start with a forward
slash (/) instead of a dash (-). To deal with this, the WSL command
conversion function will default to the original string if the WSL path
command fails.