Fixed an issue when terminal did not support color.
The Python startup file has been updated to fix an issue when the
underlying terminal does not support color codes. In this case, an error
was being generated. The startup file now gracefully handles this
condition. Additionally, the code for getting color codes was moved
into a helper function and the PS1 prompt helper function was removed.