Aligned Python's PS2 prompt with its PS1 prompt.
The Python startup file has been updated so that the PS2 prompt now
aligns its text with the PS1 prompt. That is, any line on subsequent
lines will now start at the same column as the text on the first line
(excluding the PS1 prompt naturally). This is done so function
definitions look nicer.