Changed from using the tab stop variable to the soft tab stop.
The Vimrc has been updated so that the soft tab stop variable is used to
govern how many spaces a tab key inserts when it is pressed. When the
tab stop variable is changed, this changes how many spaces a tab
character represents, changing how the file looks.

Using soft tab stop instead controls how many spaces a tab key press
represents, while preserving the original tabs. It also has the nice
side effect of allowing backspacing to be done at the tab stop
granularity for leading white space. This option was changed so that it
is clear in a file where tabs are used versus whitespace.