Added in incremental search and auto change directory.
Two new options have been added to the Vimrc file. The first is
incremental search, which displays the results of the search string as
it is typed, similar to how a web browser behaves. The second is auto
change directory, which changes the working directory to the directory
of the file currently being edited.