Changed the default spell checking behavior.
The Vimrc has been updated so that the rules for using spellchecking are
now inclusion based. Before, all files were automatically spellchecked
and auto commands defined which were not. Now, auto command define which
are spellchecked. The reason for this is because Vim's syntax rules for
various file types are incomplete, so irrelevant parts of the code can
be spellchecked. Also, some files may contain many errors, so it's nice
not to be bombarded with them when opening a file.

In addition, the spellcheck shortcut was extended so that it could be
used from insert mode.