Removed the default over length setting from the Vimrc.
The over length highlighting, used when a line exceeds the text width,
is no longer enabled by default with an auto command. Instead, new
mapping have been added to enable and disable this highlighting. This is
turned off now because some files may have a lot of over length lines,
so it's nice not to be bombarded with the red highlighting when opening
a file.