Added the Highlight MatchIt and Superman plugins to the Vimrc.
The Highlight MatchIt plugin adds highlighting of matching keywords that
are matched by the MatchIt plugin. The SuperMan plugin exposes a `vman`
command that allows for Unix Man pages to be viewed inside Vim. This
support already exists natively in Vim; this plugin simply exposes it as
an external command. The Bashrc was also updated appropriately for this
new command.