Fixed an incorrect use of set in auto commands.
If a variable is a global variable, then use set instead of set local
will cause it to change the variable for all buffers. This is an issue
for auto commands, because this can change global options that only
should be changed for the local buffer. The set local command is now
used instead.