Changed the FZF command to use RipGrep whenever possible
The default FZF command and the command used for the Control-T shortcut
have been changed to use RipGrep to list the files, instead of GNU Find.
Additionally, standard lists of files to ignore for classes of commands
were added to the Bashrc. Also, an issue with exclude patterns not being
properly quoted for the Rsync alias was resolved.