Added Bash aliases for FZF and Ripgrep.
Several aliases and functions were added to the Bashrc for FZF and
Ripgrep. Some of these set default options for the commands, while other
provide versions of the commands that only search files that are not
ignored by Git. Additionally the FZF default command variable was
removed, and trivial errors from Ripgrep are now ignored.