Added the Vimpager command to both Bash and Vim.
The Vimpager plugin has been been added to the Vimrc. This adds support
for a "less mode" to Vim, which allows it to be properly used as a
pager. Installing the plugin also installs both the Vimpager and Vimcat
commands. The Bashrc has been updated to include paths to these commands
and their manual pages.