Added the concept of a directory for user-local installed programs.
The Bashrc now defines a standard directory where all user-local
installed programs and their related files (e.g. manual pages) should be
placed. This is the standard `~/.local` path, and is used to prevent a
new path entry from having to be added for every new program that's
installed locally.

The Vim plugins that generate command line programs were also updated to
place their programs in this directory when possible (FZF is an
exception to this). Additionally, the Go path was moved to the path
variables sections, and the different types of paths are now
distinguished in comments in the Bashrc.