Added a wild ignore list to Vim.
The wild ignore list defines a list of glob patterns to ignore when
completing file and directory names (and also affects globbing functions
within Vim). This list is exported from the Bashrc file. A list of glob
patterns to ignore for text editing was added to the Bashrc.
Additionally, the existing ignore lists had their naming cleaned up.