Added in local, machine-specific settings to the Vimrc.
The Vimrc file now supports including local settings. These are handled
by a conditional sourcing of the file. If the file does not exist, then
it is not sourced. The main purpose of the local settings is to provide
settings that are specific to the local machine, and don't necessarily
belong in the main Vimrc file.