Fixed a few issues with automatically installed Vim-Plug.
The Vimrc file will automatically download the Vim-Plug file if it is
not present in the auto load directory. However, it was incorrectly
using a defined variable. Additionally, the script no longer tries to
setup the plugins, as this requires a restart of Vim. Instead, it simply
reports to the user that Vim must be restarted and exits the script.