Separated the various Vim editor files into separate directories.
The Vimrc has been updated so that the various editor files (e.g. swap
files, undo history files, etc.) are no in separate subdirectories. This
is done to just clean up the structure a bit. Also, the view directory
was changed from its default to also be in this directory.