Added support for automatically saving and restoring view to Vim.
The Vimrc has been updated to include auto commands that automatically
save and restore views whenever a buffer is exited and entered,
respectively. Views are Vim's way of keeping context about files being
editor, and they store information such as the cursor position and any
folds in the file.

Also, the Git ignore file was updated to ignore the editor files
directory, where all of the generated files from Vim are placed.