Disabled whitespace cleanup for Git commit and related files.
Whitespace cleanup was disabled for Git commit, Git send email, and diff
(patch) files. All of these files contain Diffs that may contain the
patterns that would be cleaned up by these commands. Naturally, the
original content of these Diffs shouldn't be modified, so the commands
are disabled.

Also, the event to set buffer-local variables was changed to
`BufReadPre`. This was done because `BufRead` is triggered after all of
the Vim files are sourced for the given file type, so it won't be able
to override the variable.