Disabled the saving of options for views in Vim.
The Vimrc has been updated so that the "options" are no longer saved
with views. The options include both the options for the current buffer
and any mappings that have been defined. Saving options is problematic
for several cases.

For example, if a file is opened in read-only mode, then when exiting
the file, the `buftype` option will remain as read-only, make the file
unsuitable for the next time it is opened (which can be resolved by
manually overwriting it). Additionally, this causes some options to
become "stuck", as their values will overwrite in whatever is in
configuration files.  # Please enter the commit message for your
changes. Lines starting