Decreased Vim's update time to 100 ms.
The update time often controls how long after no keys were pressed for
Vim to save the swap file to disk. Notably, this is used by the Git
gutter plugin to control how often the status of lines are updated. This
was reduced to give it a more real-time like behavior.