Changed the FZF command and shortcuts to use FD.
The file finder (FD) program is one that is in a similar vein as
Ripgrep, except that it exacts as a replacement for Find instead of
Grep. The FZF command and shortcuts have been updated to use this
command instead, as it is specifically designed for locating files.
Also, rather than setting an alias, the FZF default command environment
variable is now set instead.