Added reasonable default options for the FZF commands to the Bashrc.
The Bashrc has been updated to set default options for the FZF command
and shortcuts. Multi-select mode was enabled for the CTRL-T shortcut and
standard command. The CTRL-R, or history search, shortcut was updated to
include a preview of the command being selected. This was done to deal
with commands that don't fit in the terminal screen, as FZF doesn't
support horizontal scrolling.