Added some aliases for the FD command to the Bashrc.
The Bashrc has been updated with new aliases for the FD command. The
command was given default options so that it would skip as few files as
possible. Also, a Git variant of the FD command was added, which will
perform the file search from the root of the Git repository, if the
current working directory is inside of one.