Fixed an issue with the Ripgrep and FD ignore patterns in the Bashrc.
The issue was that the single quotes specified around the globbing
patterns were causing issues when they were used as the arguments to
commands. Thus, these are removed, and were not needed, as names with
whitespace aren't supported at the moment anyway. Also, the Ripgrep Git
function had missing ignore patterns added to it.