Changed Git's default push method from simple to current.
The Git configuration has been updated so that the default push method
is current now. This is similar to simple, but it pushes to the remote
branch with the same name by default. If the remote has no such branch,
then it will automatically create the remote branch, skipping the need
for the `-u` flag.