Added some improvements for XML files to Vim.
The first is that NuGet configuration files are now picked up as XML
files, as they are in XML format. Additionally, the highlighting from
the MatchUp plugin has been disabled for XML files. This was done
because it highlights the entire interior of tags, including tags that
only span a single line.