Added in the use of the Stow program for both Vim and Bash.
Stow is a program that acts as a basic package manager for installations
of manually installed software, or software built from source. It's
fairly bare-bones, implementing this management through symbolic links
to a defined directory where packages are placed.

To this end, the directory for these packages has been defined in the
Bashrc file. Additionally, the Vim pager and superman plugins, which
install programs from source, have been updated to use stow to install
their programs.