Updated the notify functions for Bash.
The Bashrc has been updated with new and improved functions for running
a command and notifying when it has completed, along with its exit
status. One function creates a popup window when the command completes,
while the other creates a desktop (toast) notification. Also, an issue
dealing with spaces in the commands and arguments was fixed.