Updated the CTRL-T shortcut for FZF to also include directories.
The Bashrc has been updated so that the default command for the CTRL-T
shortcut (used to do a fuzzy search to insert arguments into a command)
now also includes directories. This was done because it's useful to
include directories for arguments, and all other FZF shortcuts only do
files or directories exclusively.