Added handling of aliases for the notify function in Bash.
The Bashrc has been updated so that the notification functions, which
runs a command then notifies when it completes, can now handle aliases
passed to the function. Inside of function, Bash does not expand
aliases. Thus, to emulate this behavior, the `eval` built-in is used to
expand the alias when the command is run (if the command is not an alias
then it will remain unchanged).