Updated Bash's FZF edit shortcut to use the full FD command.
The FZF edit shortcut now utilizes the full FD command to perform the
search, rather than relying the default FZF command. This was done
because the default command now deals with both directories and files,
while this shortcut is only interested in files. Also, the command now
properly deals with spaces in file names.