Disabled the setting of the pipe fail option for Bash.
The Bashrc has been updated so that it no longer sets the pipe fail
option, which sets the exit code of the pipeline to the rightmost
command in the pipe with a non-zero exit status. This option is really
only useful in scripts, and doesn't provide much value in an interactive
prompt.

Additionally, this causes issues with some of the FZF keybindings. Since
the FD command (as does find) returns a non-zero exit status when it
does not perform a complete search, this causes the FZF CD shortcut to
only work after all directories have been enumerated. If FD's output is
cut short, and it doesn't complete the entire search, then it returns a
non-zero exit status, indicating the whole pipeline has failed.