Added the Gutentags plugin to Vim.
The Gutentags plugin is a manager for Ctags files. The plugin acts a
seamless and transparent manager of tags files. It handles generating
tags files for projects (defined by repository boundaries by default)
and incrementally updating the tags file as new changes are saved. This
is all handled in the background.

Also, good default settings were added for the plugin.