Added shortcuts for fuzzy searching files with FZF to Vim.
The Vimrc has been updated to add two new keyboard shortcuts to Vim for
fuzzy file searching. The first shortcut, CTRL-P, is modeled off of the
Sublime shortcut, and performs a fuzzy file search from the project root
directory if the current file is in a project (e.g. Git repository), or
from the current working directory if not. The second shortcut, CTRL-G
CTRL-P, performs the same search, except is restricts it to files
tracked by Git.

These shortcuts rely on the Gutentags project finder.