Added the ability to use FZF for reverse history search in Python.
The Python startup file has been updated so that there is now an
alternate shortcut, ALT-R, for performing a reverse history search. This
utilizes the FZF program to search the history, with identical options
as in Bash.

Unfortunately, there isn't a clean way to hook directly into the
Readline library with a Python function, so best the function can do is
to print the command to standard output.