Added proper support for FZF history search in Python.
The Python startup has been updated so that the FZF history search is
more reasonably supported. The new method uses a bit of a roundabout way
to achieve the desired behavior.

First, the history search shortcut invokes a Python function to perform
the search. This function sets a global variable (which is an empty
string all other times). Next, a pre-hook function has been added, which
runs after the prompt is printed, but before the user types any input.
This function inserts the selected history line into the line buffer,
making it as if the user typed the string.

The net result is that a single line is printed with the invocation of
the FZF history search function, but the next line contains the result
of the search, which is much more preferable to the old method. Also,
this has been setup to override the default history search.