Fixed the automatic handling of binary files for Vim's Hexmode plugin.
The automatic binary file detection logic added in the Vimrc was missing
a few key cases. One case was when vim is used to view a file produced
by process substitution. This is a stranger case as the file simply
appears as a symbolic link to an invalid file. The other missing case
was for empty files. Last, the buffer new file event was removed as a
trigger for the detection, as not yet created files will not be binary.