Enabled the displaying of indicator characters for whitespace to Vim.
The `list` and `listchars` options for Vim have now been enabled. The
purpose of these options is to display indicator characters for various
types of whitespace characters and certain situations related to the
editor. For example, this controls if indicator characters are used to
denote where tab characters are or if a line continues past the screen
when line wrapping is disabled. There are a few other cases that it
handles as well.