Changed Vim to not wrap lines in the middle of a word.
By default, when wrapping long lines that exceed the width of the
screen, Vim will break the line at the last character that fits on the
screen, which can cause a break to occur in the middle of a word. Thus,
to improve readability, especially with the show break option, the
settings were changed so that Vim will only break lines along word
boundaries, as defined by certain characters.