Made the run Windows command function for Bash more robust.
Because insider builds have a tendency to break the functionality of WSL
path, the run Windows command function has been updated to handle when
the utility fails. In that case, the argument is not converted and is
simply left in its original form. The recent change that led to this fix
was that the WSL path utility would fail in the argument was a
non-existent path, but only when converting to Windows paths.