Created a proper fix to make the run Windows command function robust.
The attempt in the earlier PR was both insufficient and changed the
wrong parts of the Bash function to run a Windows command. All that was
needed instead was defaulting to the original input element for the
command being run, similar to how the arguments are treated. Thus, when
WSL path fails, it acts simply as an identity.