Revert "Changed Git's default push method from simple to current."
This reverts commit 9a7d5f0623dfc944356f8fcf3259f9b4ecc7daac.

This commit was reverted because the current setting for push.default is
insufficient. In particular, while it will automatically create an
upstream branch, it does not create tracking information for Git pull.
Thus, the upstream branch must be set manually anyway. To prevent this
annoying behavior, it's easier to simply set the tracking information
when creating the upstream branch with the `-u` option for Git push.