Migrated the history settings and shell options out of the Bashrc file.
The history settings and shell options for Bash were moved out of the
main Bashrc file, into their own files in a Bash directory. This is the
first step in splitting up the Bashrc file into smaller files dedicated
groups of settings.

As part of these changes, the Gitignore was updated to allow for Bash
files in the Bash directory and the Bashrc was updated to source these
newly created files.