Changed the verbosity of the Git commit command.
The Git configuration file was updated to decrease the verbosity of the
Git commit command. Previously, when the editor was opened up, the diff
of all staged and unstaged changes was displayed at the bottom of the
file. Since seeing the unstaged changes is not as useful, the verbosity
has been decreased to only display the diff of staged changes.