Updated Bash's command prompt to allow for a user added prefix.
The PS1 setup for bash now supports having a user-defined prefix to the
prompt, which comes via the environment variable `PS1_PREFIX`. This can
be used by parent shell to set a prefix for child subshells that they
launch. This is useful when context-based environments within in the
shell need to be distinguished. For example, this is used by WSL aliases
for launching shells that have a Windows CoreXT enlistment to a specific
repository.