Added a script for creating symbolic links to the Dotfiles.
A script has been added to the repository that handles setting up a
system to utilize the configuration files in the Dotfiles repository.
This script simply creates symbolic links in the user's home directory
to appropriate configuration files in the repository. These are the
locations of standard system configuration files (e.g. Bashrc, Git
configuration, Vim directory, etc.). This ensures that programs will
utilize the configuration files in this repository.