Updated the creation of symbolic links to gracefully handle directories.
The symbolic link script has been updated so that when a symbolic link
has to overwrite a preexisting directory, it will gracefully do so. The
standard Ln command will not overwrite a directory will a symbolic link;
it will either refuse to do or it will place the link file inside the
directory. Thus, it must be manually done.

The directories are now simply backed up in the same manner as other
files are with Ln, with the `.~<n>~` prefix, where n is some number.