Fixed an issue with how the script directory is determined.
Respecting symbolic links in the current working directory is important
so that relative paths can be established for the configuration files.
This matters when the path to the script is relative, as the real path
utility will use the true path, even with the no symlinks option
specified. Thus, the path is converted to an absolute path by prepending
the current working directory. For absolute paths, this is irrelevant,
so the normal method can be used.