Fixed how items are added to path variables in the Bashrc.
Previously, paths were already appended to the existing path variable
(e.g. `PATH`, `MANPATH`, `LD_LIBRARY_PATH`, etc.). However, the
precedence for path variables is left to right. That is, if a match is
found in an earlier entry in the variable, that will be used over a
later one. Thus, system paths can (and did) override the versions of
executables and other artifacts installed in local or user-defined
directories. Instead, paths are now prepended to path variables to
prevent this.