Cleaned up how path variables are defined in the Bashrc.
The prepending of values has been cleaned up so spurious colons are not
generated in any of the path variables. When prepending to a path
variable, a special construct is used so that a colon is added for the
separator only if the path variable already contains values. This only
needs to be done the first time the variable is defined in the Bashrc,
so it has been done only there.