Standardized all external Git aliases to use Bash, and fixed issues with summarize.
The remaining Git aliases that invoke an external shell to run have all
been standardized to use Bash. This is mainly done to have all the nice
features missing from more limited shells like Dash.

Additionally, some nuanced issues with how the summarize alias parsed
arguments was resolved. First, specifying arguments at the top level is
unnecessary as Git appends the command-line arguments to the end of an
alias. Second, because of how Bash is invoked, the `${0}` argument has a
different meaning than usual. This is because there is no command when
Bash is invoked, so the first argument is actually a true argument.
Thus, using common variables like "${@}" or "${#}" is insufficient
because they ignore `${0}`.