Changed how binary files are automatically detected for Vim.
When a file is opened, the previous Vim settings attempted to
automatically detect if a file was binary using the `file` utility.
If the is binary, then Vim would use the hex mode plugin to open the
binary file as a hex strings. The issue with this is that there are too
many cases to handle (e.g. only text was handled but other types JSON
were needed). Instead, the plugin's built-in mechanism for detecting
binary files based on patterns via file extensions is now used. This is
more reliable and easier to deal with.