Changed the Alt-W shortcut in the Readline configuration.
Previously, the Alt-W shortcut was set to be shell backward kill word,
which would delete the word behind the cursor, using spaces and unquoted
shell metacharacters (e.g. "$") as delimiters for words. This has been
updated to the Unix word rubout command, which does the same action, but
instead uses spaces as delimiters. In practice, this is more useful than
stopping at shell metacharacters as the Ctrl-W can be used to delete sub
space delimited word entities.