Extended the FZF edit shortcut to accept paths as arguments for Bash.
Previously, the FZF edit shortcut (Alt-e) did not accept any arguments
and would thus always search from the current working directory. Now,
the shortcut and the function that backs it accepts arguments. If a
single argument is specified, then it is treated as the root path from
which the search for files to edit is conducted. In order to be more
usable, if more than one argument is specified then it is ignored. This
is done so the user doesn't have to explicitly clear a very long
command-line.

The shortcut accomplishes accepting arguments by prefixing the current
line with usual FZF edit function with command substitution and then
ends the line with a parenthesis, so the current line is passed as
arguments into the function.