Added support for displaying the Readline editing mode in PS1.
The PS1 prompt has been updated so that it now supports displaying the
editing mode for Bash (through Readline). The default mode is Emacs, so
nothing is displayed in that case. Otherwise, for Vi editing mode,
`{vi}` is displayed in magenta color.

Also, the definition of the variable was cleaned up and each section was
split into a separate line that gets apended to the string.