Removed the showing of all completions if it is ambiguous for Readline.
Previously, Readline would immediately display all possible completion
matches if the completion was "ambiguous". "Ambiguous" specifically
means that when a completion is attempted, there are multiple matches
and a common prefix exists (e.g. the completion can fill in additional
characters). The problem is when there are hundreds of matches, Readline
will first prompt to display them and there's a significant delay to
discover them.

Now, the ambiguous completion will not trigger listing all matches, but
instead just complete the common prefix. If the user attempts the
completion again (referred to as an "unmodified" completion), then all
possible matches will be listed.