Removed the ignoring of any commands for Bash's history.
Previously, a select set of common, simple commands were ignored for
Bash's history (e.g. `cd` without any arguments). This was done to
prevent the cluttering of the Bash history file and current history
buffer. However, in practice, this turned out not to be very useful as
it makes it annoying when the previously typed command cannot be
recalled. Additionally, with tools like FZF's reverse history search,
it's very easy to recall specific commands, so cluttering is not a
concern.