Fixed poor colors for Vim's GitGutter plugin.
A recent update to the GitGutter plugin stopped explicitly setting the
colors, and instead had the plugin attempt to derive appropriate colors
based on the color scheme. However, this didn't work very well as the
default color scheme's sign column has a horrible gray background that
obscures most text colors.

Thus, the plugin colors were updated manually to be better colors.
Additionally, instead of simply reverting back to white as the
foreground color, green is used for additions, yellow for changes, and
red for deletions.

This also required some rearranging of where the Plugins Vim file is
sourced, as it needs to happen after the color scheme is set.